# openapi/sgf-member-v1.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "SGF private member bridge",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:42837/"
    }
  ],
  "paths": {
    "/api/v1/member/avatar": {
      "post": {
        "tags": [
          "MemberAvatar"
        ],
        "operationId": "Member_AvatarUpload",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberAvatarResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberAvatarResult"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MemberAvatarResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "413": {
            "description": "Payload Too Large"
          },
          "429": {
            "description": "Too Many Requests"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/member/forgot-password": {
      "post": {
        "tags": [
          "MemberPasswordReset"
        ],
        "operationId": "Member_ForgotPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberForgotPasswordRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberForgotPasswordRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MemberForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberPasswordResetResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberPasswordResetResult"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MemberPasswordResetResult"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberPasswordResetResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberPasswordResetResult"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MemberPasswordResetResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/member/reset-password": {
      "post": {
        "tags": [
          "MemberPasswordReset"
        ],
        "operationId": "Member_ResetPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberResetPasswordRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberResetPasswordRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MemberResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberPasswordResetResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberPasswordResetResult"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MemberPasswordResetResult"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberPasswordResetResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberPasswordResetResult"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MemberPasswordResetResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/member/profile": {
      "get": {
        "tags": [
          "MemberProfile"
        ],
        "operationId": "Member_Profile",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberProfileEditDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberProfileEditDto"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MemberProfileEditDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      },
      "post": {
        "tags": [
          "MemberProfile"
        ],
        "operationId": "Member_ProfileUpdate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberProfileEditRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberProfileEditRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MemberProfileEditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberProfileEditResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberProfileEditResult"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MemberProfileEditResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberProfileEditResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberProfileEditResult"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MemberProfileEditResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/member/register": {
      "post": {
        "tags": [
          "MemberRegistration"
        ],
        "operationId": "Member_Register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberRegistrationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberRegistrationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MemberRegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberRegistrationResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberRegistrationResult"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MemberRegistrationResult"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberRegistrationResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberRegistrationResult"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MemberRegistrationResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/member/login": {
      "post": {
        "tags": [
          "MemberSession"
        ],
        "operationId": "Member_Login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberLoginRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberLoginRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MemberLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberLoginResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberLoginResult"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MemberLoginResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/member/session": {
      "get": {
        "tags": [
          "MemberSession"
        ],
        "operationId": "Member_Session",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberSessionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberSessionDto"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MemberSessionDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/member/logout": {
      "post": {
        "tags": [
          "MemberSession"
        ],
        "operationId": "Member_Logout",
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
      "MemberAvatarResult": {
        "required": [
          "profileImageUrl"
        ],
        "type": "object",
        "properties": {
          "profileImageUrl": {
            "type": "string"
          }
        }
      },
      "MemberForgotPasswordRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
      "MemberLoginRequest": {
        "required": [
          "username",
          "password"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": [
              "null",
              "string"
            ]
          },
          "password": {
            "type": [
              "null",
              "string"
            ]
          },
          "rememberMe": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "MemberLoginResult": {
        "required": [
          "succeeded"
        ],
        "type": "object",
        "properties": {
          "succeeded": {
            "type": "boolean"
          }
        }
      },
      "MemberPasswordResetResult": {
        "required": [
          "succeeded",
          "errors"
        ],
        "type": "object",
        "properties": {
          "succeeded": {
            "type": "boolean"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "MemberProfileChoice": {
        "required": [
          "key",
          "name"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "MemberProfileEditDto": {
        "required": [
          "values",
          "profileImageUrl",
          "skills",
          "groups"
        ],
        "type": "object",
        "properties": {
          "values": {
            "$ref": "#/components/schemas/MemberProfileEditRequest"
          },
          "profileImageUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberProfileChoice"
            }
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberProfileChoice"
            }
          }
        }
      },
      "MemberProfileEditRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "firstName": {
            "type": [
              "null",
              "string"
            ]
          },
          "lastName": {
            "type": [
              "null",
              "string"
            ]
          },
          "jobTitle": {
            "type": [
              "null",
              "string"
            ]
          },
          "aboutText": {
            "type": [
              "null",
              "string"
            ]
          },
          "city": {
            "type": [
              "null",
              "string"
            ]
          },
          "state": {
            "type": [
              "null",
              "string"
            ]
          },
          "availableForHire": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "availableForContractWork": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "twitterUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "twitchUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "facebookUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "instagramUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "linkedInUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "meetupUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "websiteUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "youTubeUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "skills": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "groups": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
      "MemberProfileEditResult": {
        "required": [
          "succeeded",
          "errors"
        ],
        "type": "object",
        "properties": {
          "succeeded": {
            "type": "boolean"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "MemberRegistrationRequest": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": [
              "null",
              "string"
            ]
          },
          "lastName": {
            "type": [
              "null",
              "string"
            ]
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "username": {
            "type": [
              "null",
              "string"
            ]
          },
          "password": {
            "type": [
              "null",
              "string"
            ]
          },
          "challengeQuestion": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "MemberRegistrationResult": {
        "required": [
          "succeeded",
          "errors"
        ],
        "type": "object",
        "properties": {
          "succeeded": {
            "type": "boolean"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "MemberResetPasswordRequest": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          }
        }
      },
      "MemberSessionDto": {
        "required": [
          "username",
          "name"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "MemberAvatar"
    },
    {
      "name": "MemberPasswordReset"
    },
    {
      "name": "MemberProfile"
    },
    {
      "name": "MemberRegistration"
    },
    {
      "name": "MemberSession"
    }
  ]
}